Add Hangfire.JobsLogger.ExampleShared
