+++ src/samples/Hangfire.JobsLogger.ExampleCore/Hangfire.JobsLogger.ExampleCore.csproj
@@ -18,6 +18,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\main\Hangfire.JobsLogger.csproj" />
+    <ProjectReference Include="..\Hangfire.JobsLogger.ExampleShared\Hangfire.JobsLogger.ExampleShared.csproj" />
   </ItemGroup>
 
   <ItemGroup>
