+++ src/samples/Hangfire.JobsLogger.ExampleCore/Startup.cs
@@ -1,9 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using Hangfire.JobsLogger.ExampleShared;
 using Hangfire.LiteDB;
-using Hangfire.Server;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
@@ -44,17 +40,9 @@ namespace Hangfire.JobsLogger.ExampleCore
                 await context.Response.WriteAsync("Hello World!");
             });
 
-            RecurringJob.AddOrUpdate(() => TaskMethod(null), Cron.Minutely);
-        }
+            var taskExample = new TaskExample();
 
-        public void TaskMethod(PerformContext context)
-        {
-            context.LogTrace($"Trace Message.. {DateTime.UtcNow.Ticks}");
-            context.LogDebug($"Debug Message.. {DateTime.UtcNow.Ticks}");
-            context.LogInformation($"Information Message.. {DateTime.UtcNow.Ticks}");
-            context.LogWarning($"Warning Message.. {DateTime.UtcNow.Ticks}");
-            context.LogError($"Error Message.. {DateTime.UtcNow.Ticks}");
-            context.LogCritical($"Critical Message.. {DateTime.UtcNow.Ticks}");
+            RecurringJob.AddOrUpdate(() => taskExample.TaskMethod(null), Cron.Minutely);
         }
     }
 }
