+++ src/samples/Hangfire.JobsLogger.ExampleNET/Hangfire.JobsLogger.ExampleNET.csproj
@@ -133,6 +133,10 @@
       <Project>{7a934ab8-8f4d-40a4-bd3f-98e9914ecac6}</Project>
       <Name>Hangfire.JobsLogger</Name>
     </ProjectReference>
+    <ProjectReference Include="..\Hangfire.JobsLogger.ExampleShared\Hangfire.JobsLogger.ExampleShared.csproj">
+      <Project>{282e15fe-a782-4a5c-ae78-5d7d7bc5d0a8}</Project>
+      <Name>Hangfire.JobsLogger.ExampleShared</Name>
+    </ProjectReference>
   </ItemGroup>
   <PropertyGroup>
     <VisualStudioVersion Condition="'$(VisualStudioVersion)' == ''">10.0</VisualStudioVersion>
